mcp-client/pyproject.toml
[project]
name = "weather-client"
version = "0.1.0"
description = "HTTP Weather Client with Ollama"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "python-dotenv>=1.0.1",
]
